Moved CriteriaType enum out of Criteria class and into own file
