+++ src/main/java/com/microsoft/azure/spring/data/documentdb/core/DocumentDbTemplate.java
@@ -10,7 +10,7 @@ import com.microsoft.azure.documentdb.*;
 import com.microsoft.azure.documentdb.internal.HttpConstants;
 import com.microsoft.azure.spring.data.documentdb.DocumentDbFactory;
 import com.microsoft.azure.spring.data.documentdb.core.convert.MappingDocumentDbConverter;
-import com.microsoft.azure.spring.data.documentdb.core.query.Criteria.CriteriaType;
+import com.microsoft.azure.spring.data.documentdb.core.query.CriteriaType;
 import com.microsoft.azure.spring.data.documentdb.core.query.Criteria;
 import com.microsoft.azure.spring.data.documentdb.core.query.Query;
 import com.microsoft.azure.spring.data.documentdb.repository.support.DocumentDbEntityInformation;
